Json deserializer collection perf improvements by avoiding boxing (#36865)
