+++ src/System.Text.Json/tests/System.Text.Json.Tests.csproj
@@ -40,6 +40,9 @@
     <Compile Include="Serialization\Stream.WriteTests.cs" />
     <Compile Include="Serialization\TestClasses.cs" />
     <Compile Include="Serialization\TestClasses.Polymorphic.cs" />
+    <Compile Include="Serialization\TestClasses.SimpleTestClass.cs" />
+    <Compile Include="Serialization\TestClasses.SimpleTestClassWithNullables.cs" />
+    <Compile Include="Serialization\TestClasses.SimpleTestClassWithObject.cs" />
     <Compile Include="Serialization\TestData.cs" />
     <Compile Include="Serialization\Value.ReadTests.cs" />
     <Compile Include="Serialization\Value.WriteTests.cs" />
